+++ code/MyEncoder/MyEncoder.h
@@ -19,8 +19,8 @@ typedef struct
 #define Switch_ENCODER_L         TIM3_ENCODER_CH1_P02_6        //右编码器计数引脚
 #define Switch_ENCODER_R         TIM3_ENCODER_CH2_P02_7        //右编码器方向引脚
 
-extern int16 Encoder_speed_l;
-extern int16 Encoder_speed_r;
+extern int Encoder_speed_l;
+extern int Encoder_speed_r;
 extern int switch_encoder_num;
 extern int switch_encoder_change_num ;
 extern uint8 switch_encode_bring_flag;
